fix imdb problem with subquery on actors
Using imdb_ref : [] raises the same error as it was before at the movies because there are imdb-id with a \n at the end.
Using key : [{namespace: ..., value: None}] solves the error, but the resultset is much smaller (7 vs 200k), so I ignore
this field and get all actors of a film wether they have a link to imdb or not.